Revert "Windows Support"
