+++ .github/workflows/go.yml
@@ -25,7 +25,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [macOS-latest,ubuntu-latest,windows-latest]
+        os: [macOS-latest,ubuntu-latest]
     steps:
     - name: Set up Go 1.16
       uses: actions/setup-go@v1
