+++ benchmarks/eviop-echo-server/main.go
@@ -3,7 +3,6 @@
 
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
-// +build !windows
 
 package main
