+++ benchmarks/gev-echo-server/echo.go
@@ -6,6 +6,8 @@ import (
 	_ "net/http/pprof"
 	"strconv"
 
+	"github.com/Allenxuxu/gev/log"
+
 	"github.com/Allenxuxu/gev"
 )
 
@@ -31,6 +33,7 @@ func (s *example) OnClose(c *gev.Connection) {
 }
 
 func main() {
+	log.SetLevel(log.LevelDebug)
 	go func() {
 		if err := http.ListenAndServe(":6089", nil); err != nil {
 			panic(err)
