+++ benchmarks/websocket/server.go
@@ -4,6 +4,8 @@ import (
 	"flag"
 	"strconv"
 
+	"github.com/Allenxuxu/gev/log"
+
 	"github.com/Allenxuxu/gev"
 	"github.com/Allenxuxu/gev/plugins/websocket"
 	"github.com/Allenxuxu/gev/plugins/websocket/ws"
@@ -39,6 +41,7 @@ func NewWebSocketServer(handler websocket.WSHandler, u *ws.Upgrader, opts ...gev
 }
 
 func main() {
+	log.SetLevel(log.LevelDebug)
 	var (
 		port  int
 		loops int
