+++ context_test.go
@@ -1,6 +1,7 @@
 package gev
 
 import (
+	"fmt"
 	"testing"
 )
 
@@ -17,24 +18,24 @@ func TestKeyValueContext(t *testing.T) {
 
 	// Get
 	if i, ok := ctx.Get("1"); !ok || i.(int) != 1 {
-		t.Fatalf("i should be %v, but %v", i, i)
+		t.Fatal(fmt.Sprintf("i should be %v, but %v", i, i))
 	}
 	if i, ok := ctx.Get("2"); !ok || i.(int) != 2 {
-		t.Fatalf("i should be %v, but %v", i, i)
+		t.Fatal(fmt.Sprintf("i should be %v, but %v", i, i))
 	}
 	if i, ok := ctx.Get("3"); !ok || i.(int) != 3 {
-		t.Fatalf("i should be %v, but %v", i, i)
+		t.Fatal(fmt.Sprintf("i should be %v, but %v", i, i))
 	}
 
 	// Delete
 	ctx.Delete("1")
 	if _, ok := ctx.Get("1"); ok {
-		t.Fatalf("ok should be false, but %t", ok)
+		t.Fatal(fmt.Sprintf("ok should be false, but %t", ok))
 	}
 
 	// reset
 	ctx.reset()
 	if _, ok := ctx.Get("2"); ok {
-		t.Fatalf("ok should be false, but %t", ok)
+		t.Fatal(fmt.Sprintf("ok should be false, but %t", ok))
 	}
 }
