+++ eventloop/eventloop_test.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package eventloop
 
 import (
