+++ poller/poller_test.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package poller
 
 import (
