+++ server.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package gev
 
 import (
