+++ server_conn_test.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package gev
 
 import (
@@ -7,6 +5,8 @@ import (
 	"net"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
 
 	"github.com/Allenxuxu/toolkit/sync"
 
@@ -34,6 +34,7 @@ func (s *example2) OnClose(c *Connection) {
 }
 
 func TestConnClose(t *testing.T) {
+	log.SetLevel(log.LevelDebug)
 	handler := new(example2)
 
 	s, err := NewServer(handler,
@@ -79,6 +80,7 @@ func (s *example3) OnClose(c *Connection) {
 	// log.Info("OnClose")
 }
 func TestIdleTime(t *testing.T) {
+	log.SetLevel(log.LevelDebug)
 	handler := new(example3)
 
 	s, err := NewServer(handler,
@@ -95,17 +97,13 @@ func TestIdleTime(t *testing.T) {
 
 	start := time.Now()
 	wg := &sync.WaitGroupWrapper{}
-	log.Infof("Server start")
-
 	for i := 0; i < 100; i++ {
 		wg.AddAndRun(func() {
-			conn, err := net.DialTimeout("tcp", "127.0.0.1:1830", time.Second*3)
+			conn, err := net.DialTimeout("tcp", "127.0.0.1:1830", time.Second*60)
 			if err != nil {
 				log.Error(err)
 				return
 			}
-
-			log.Infof("Client conn success %v", conn.LocalAddr())
 
 			buf := make([]byte, 10)
 			n, err := conn.Read(buf)
@@ -119,6 +117,72 @@ func TestIdleTime(t *testing.T) {
 	et := time.Since(start)
 	if et.Seconds() > 4 || et.Seconds() < 3 {
 		t.Fatal(et.Seconds())
+	}
+
+	s.Stop()
+}
+
+func TestConnLoadBalanceLeastConnection(t *testing.T) {
+	handler := new(example3)
+
+	s, err := NewServer(handler,
+		Network("tcp"),
+		Address(":1840"),
+		NumLoops(4),
+		ReusePort(true),
+		LoadBalance(LeastConnection()))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	go s.Start()
+
+	for i := 0; i < 200; i++ {
+		_, err := net.DialTimeout("tcp", "127.0.0.1:1840", time.Second*60)
+		if err != nil {
+			t.Fatal(err)
+		}
+		time.Sleep(time.Millisecond * 20)
+	}
+
+	time.Sleep(time.Millisecond * 20)
+
+	for i := 0; i < len(s.workLoops); i++ {
+		assert.Equal(t, int64(50), s.workLoops[i].ConnectionCount())
+	}
+
+	s.Stop()
+}
+
+func TestConnLoadBalanceRoundRobin(t *testing.T) {
+	handler := new(example3)
+
+	s, err := NewServer(handler,
+		Network("tcp"),
+		Address(":1841"),
+		NumLoops(4),
+		ReusePort(true),
+		LoadBalance(RoundRobin()))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	go s.Start()
+
+	for i := 0; i < 9; i++ {
+		_, err := net.DialTimeout("tcp", "127.0.0.1:1841", time.Second*60)
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+	time.Sleep(time.Millisecond * 20)
+
+	for i := 0; i < len(s.workLoops); i++ {
+		expected := 2
+		if i == 0 {
+			expected = 3
+		}
+		assert.Equal(t, expected, int(s.workLoops[i].ConnectionCount()))
 	}
 
 	s.Stop()
